feat: Add logic to change the puesto in the programa operacional
This commit adds logic to change the puesto in the programa operacional based on the received message. It retrieves the programa operacional using the NumeroAZ from the message and updates the puesto value. If the programa operacional is not found, a warning message is logged. This logic is added to the Consume method in the RegistrarMarcaAtraqueConsumer class.

Co-authored-by: [Author Name]