refactor: Update IClienteRepository and ClienteRepository
This commit updates the IClienteRepository interface and the ClienteRepository class in the SiopjModule.Domain.Repositories and SiopjModule.Infraestructure.Database.Repository namespaces, respectively. The changes include adding a new method, UltimoCodigo(), to the IClienteRepository interface and implementing it in the ClienteRepository class. The UltimoCodigo() method retrieves the maximum value of the Codigo property from the Cliente entity where the Codigo is less than 9000. This update enhances the functionality of the repository by providing a way to get the last used code for a Cliente.

Co-authored-by: [Author Name]